Fixed issue with MODESDICT mapping both lss_l and lss_m to tech of LSS,LM.
This resulted in the code running for an open filter, but producing the wrong band, but crashed when trying to apply a filter, as the wavelength coverage became zero.

Fixed by using the "mode" line in the recipes.yaml file to set the mode directly, and adding a -m --mode option to the arg parser for the raw_script.py command line module. We may want add some error trapping to catch inconsistent mode/tech keywords.