Some major updates to the calling of run_recipes.
MJDOBS Determination
--------------------

The date of observation is now determined as follows

 - if --sequence is set to a date in the format yyyy-mm-dd hh:mm:ss
   the code will start from that date and increment based on
   observing time of the previous call to simulate.

 - if --sequence = 1, the observations will start from the date
   given in the first entry in the YAML file

 - if --sequence is not set, the observations will take the date
   directly from the YAML file. If an entry other than the first one
   doesn't have mjdobs in the YAML entry, the observation time will
   be incremented as above.

mjdObs is no longer an expandable.

Instead, there is a new paramter, nObs, in the YAML file. nObs observations
of each combination in the entry will be generated with sequential timestamps,
based either on the previous exposure or the mjdObs value depending
on the --sequence options.

e.g.

>

CALIBRATION FILES
-----------------

If the option --doCalib is set to an integer > 0, the script will determine
what combination of darks and flats need to be observed as part of the sequence of
observations, and will execute them at the end of the observations given in the
YAML file.

At the moment, DARK,WCUOFF files are not included; they may be added later.

The value of doCalib, if > 0, determines the number of each type to be
execute.

DEBUGGING
---------

Some more informative output is written to the screen; the values of the
run time options, and the parameters of each call to simulate().

If the command line option --testRun is set, the script will be run,
the YAML file parsed, and mdjObs calculated, but simulate() will not
be called. This is intended for debugging and checking input.

Added a routine checking the YAML file for valid input. It will check that
mode, prefix, properties, dit, ndit, filter_name, catg, tech, type, nObs exist,
check that ndit and nObs are positive integers, dit a positive number,
and for valid values of mode, catg, tech, type (as imported from simulationDefinitions).
If invalid input is detected, an error message is written and the code aborts.

Currently filter_name, ndfilter_name or source are not checked.

PREFIX
------

This was set before, but to highlight the change. Each YAML entry now
has a prefix field in addition to mode, because the labels for each
recipe need to be unique, while the prefix to the output file does not.
prefix is solely used to label the entries; mode is used for generating
file names and setting DO.CATG.